move NativeType to abstract
