+++ tests/test_transpiler_star_connectivity.py
@@ -66,7 +66,7 @@ def test_fix_connectivity_unitaries(nqubits, unitary_dim, depth, middle_qubit):
     """Checks that the transpiled circuit can be executed and is equivalent to original
     when using unitaries."""
 
-    from .test_transpilers_unitary_decompositions import random_unitary
+    from .test_transpiler_unitary_decompositions import random_unitary
 
     # find the number of qubits for hardware circuit
     n_hardware_qubits = max(nqubits, middle_qubit + 1)
